+++ scripts/macros/monsterFeatures/troll/regeneration.js
@@ -1,7 +1,6 @@
 import {chris} from '../../../helperFunctions.js';
 import {constants} from '../../../constants.js';
 async function onHit(workflow, targetToken) {
-    console.log(workflow);
     if (!workflow.damageRoll || !workflow.hitTargets.has(targetToken)) return;
     let effect = chris.findEffect(targetToken.actor, 'Regeneration');
     if (!effect) return;
@@ -18,10 +17,36 @@ async function onHit(workflow, targetToken) {
     }
     if (!stopHeal) {
         if (chris.getConfiguration(originItem, 'zeroHP')) return;
-        if (targetToken.actor.system.attributes.hp.value === 0) {
-            console.log('Is defeated!');
+        if (targetToken.actor.system.attributes.hp.value === 0 && chris.inCombat()) {
+            let updates = {
+                'defeated': false
+            };
+            await chris.updateCombatant(chris.getCombatant(targetToken), updates);
+            let effectData = {
+                'icon': 'icons/svg/skull.svg',
+                'label': 'Dead?',
+                'origin': originItem.uuid,
+                'duration': {
+                    'seconds': 86400
+                },
+                'flags': {
+                    'core': {
+                        'overlay': true
+                    }
+                }
+            };
+            await warpgate.wait(200);
+            await chris.createEffect(targetToken.actor, effectData);
+            if (!chris.findEffect(targetToken.actor, 'Prone')) await chris.addCondition(targetToken.actor, 'Prone', false, originItem.uuid);
         }
         return;
+    } else {
+        let hp = targetToken.actor.system.attributes.hp.value;
+        if (hp != 0 || !chris.inCombat()) return;
+        let effect = chris.findEffect(targetToken.actor, 'Dead?');
+        if (!effect) return;
+        await chris.removeEffect(effect);
+        await chris.addCondition(targetToken.actor, 'Dead', true, null);
     }
     let effectData = {
         'label': 'Regeneration Blocked',
@@ -47,8 +72,15 @@ async function turnStart(token, origin) {
     let feature = new CONFIG.Item.documentClass(featureData, {'parent': token.actor});
     let options = constants.syntheticItemWorkflowOptions([token.document.uuid]);
     setProperty(options, 'workflowOptions.allowIncapacitated', true);
+    if (hp === 0) {
+        let effect3 = chris.findEffect(token.actor, 'Dead?');
+        if (effect3) {
+            await chris.removeEffect(effect3);
+        }
+    }
     await MidiQOL.completeItemUse(feature, {}, options);
 }
+
 export let regeneration = {
     'onHit': onHit,
     'turnStart': turnStart